hw6/pyproject.toml
[project]
name = "hw6"
version = "0.1.0"
description = ""
authors = [
    {name = "Gulnara",email = "gulnarkaz@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi (>=0.115.7,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "jinja2 (>=3.1.5,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "attrs (>=25.1.0,<26.0.0)",
    "python-jose[cryptography] (>=3.3.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
